requirements.txt
aiogram==3.13.1
aiomysql==0.2.0
virustotal-python==1.1.1
scikit-learn==1.5.2
joblib==1.4.2
nltk==3.8.2
python-dotenv==1.0.1
aiohttp==3.10.5
transformers==4.46.0